# Conflicts:
#	pkg/core/resources/apis/core/interfaces.go
#	pkg/core/resources/apis/meshmultizoneservice/api/v1alpha1/helpers.go
#	pkg/core/resources/apis/meshservice/api/v1alpha1/helpers.go
#	pkg/plugins/policies/core/rules/resolve/targetref.go
#	pkg/xds/context/context.go

+++ pkg/plugins/runtime/k8s/controllers/meshservice_controller.go
@@ -399,7 +399,7 @@ func (r *MeshServiceReconciler) manageMeshService(
 			}
 			ms.Spec.Ports = append(ms.Spec.Ports, meshservice_api.Port{
 				Name:        portName,
-				Port:        uint32(port.Port),
+				Port:        port.Port,
 				TargetPort:  port.TargetPort,
 				AppProtocol: core_mesh.Protocol(pointer.DerefOr(port.AppProtocol, "tcp")),
 			})
