+++ pkg/test/resources/builders/meshservice_builder.go
@@ -76,24 +76,24 @@ func (m *MeshServiceBuilder) WithDataplaneTagsSelectorKV(selectorKV ...string) *
 	return m.WithDataplaneTagsSelector(TagsKVToMap(selectorKV))
 }
 
-func (m *MeshServiceBuilder) AddIntPort(port, target uint32, protocol core_mesh.Protocol) *MeshServiceBuilder {
+func (m *MeshServiceBuilder) AddIntPort(port, target int32, protocol core_mesh.Protocol) *MeshServiceBuilder {
 	m.res.Spec.Ports = append(m.res.Spec.Ports, v1alpha1.Port{
 		Port: port,
 		TargetPort: intstr.IntOrString{
 			Type:   intstr.Int,
-			IntVal: int32(target),
+			IntVal: target,
 		},
 		AppProtocol: protocol,
 	})
 	return m
 }
 
-func (m *MeshServiceBuilder) AddIntPortWithName(port, target uint32, protocol core_mesh.Protocol, name string) *MeshServiceBuilder {
+func (m *MeshServiceBuilder) AddIntPortWithName(port, target int32, protocol core_mesh.Protocol, name string) *MeshServiceBuilder {
 	m.res.Spec.Ports = append(m.res.Spec.Ports, v1alpha1.Port{
 		Port: port,
 		TargetPort: intstr.IntOrString{
 			Type:   intstr.Int,
-			IntVal: int32(target),
+			IntVal: target,
 		},
 		AppProtocol: protocol,
 		Name:        name,
